refactor(src): API import/export 방식 변경 및 반영
