+++ frontend/src/hooks/template/useTemplateListQuery.ts
@@ -1,7 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { QUERY_KEY } from '@/api/queryKeys';
-import { getTemplateList } from '@/api/templates';
+import { QUERY_KEY, getTemplateList } from '@/api';
 import type { TemplateListResponse } from '@/types/template';
 
 export const useTemplateListQuery = () =>
