+++ frontend/src/hooks/template/useTemplateUploadQuery.ts
@@ -1,7 +1,6 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { QUERY_KEY } from '@/api/queryKeys';
-import { postTemplate } from '@/api/templates';
+import { QUERY_KEY, postTemplate } from '@/api';
 
 export const useTemplateUploadQuery = () => {
   const queryClient = useQueryClient();
